Resolve PR comments
